repetetive metoths extracted from the hotell calss
